% Generated by roxygen2 (4.0.1): do not edit by hand
\name{classicalModel}
\alias{classicalModel}
\title{Simulate trees}
\usage{
classicalModel(nb_diversification_shift, nb_mass_extinction, total_nb_species,
  lambda_max, lambda_min, mu_min, survival_rate_min, seed = 1)
}
\arguments{
\item{nb_diversification_shift}{fff}

\item{nb_mass_extinction}{fff}

\item{total_nb_species}{fff}

\item{lambda_max}{ff}

\item{lambda_min}{fff}

\item{mu_min}{ff}

\item{survival_rate_min}{ff}

\item{seed}{ff}
}
\description{
What it does
}
\details{
This function
}
\examples{
lol <- classical_model(1, 1, 500, 0.1, 0.01, .01, .2, seed=1506621)
}

