% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jiveMake}
\alias{jiveMake}
\title{Make jive object}
\usage{
jiveMake(simmap, traits, model.var = "OU1", model.mean = "BM",
  model.lik = "Multinorm")
}
\arguments{
\item{simmap}{an object of class "jive" (see details)}

\item{traits}{name of the output file that will store the log of MCMC chain}

\item{model.var}{sampling frequency of the MCMC chain (how often chain will be saved into output file}

\item{model.mean}{printing frequency of the MCMC chain (how often chain will be printed in the R console)}

\item{model.lik}{number of classes for thermodynamic integration (see details)}
}
\description{
Creates a jive object from intraspecific observations and species phylogeny
}
\details{
This function creates a jive object needed for \code{\link{jiveMCMC}} function.
Trait values must be stored as a matrix, where lines are vectors of observations for each species, with NA for no data.
Rownames are species names. Phylogenetic tree must be provided as either simmap object (for models with multiple regimes)
or as a phylo object (for BM or OU1 models). Rownames and tip labels of a phylogenetic tree should match exactly.
There are three models implemeted for estimation of species variances evolution - BM, OU1 and OUM. Evolution of
species means is only implemented with BM model. Species-specific distribution are models as multivariate normal distribution
}
\examples{
my.jive <- jiveMake(phy, traits,  model.var="OU1", model.mean="BM", model.lik="Multinorm")
}

