% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jiveMCMC}
\alias{jiveMCMC}
\title{Jive MCMC}
\usage{
jiveMCMC(jive, log.file = "jive_mcmc.log", sampling.freq = 1000,
  print.freq = 1000, ncat = 1, beta.param = 0.3, ngen = 5000000,
  burnin = 0, update.freq = NULL)
}
\arguments{
\item{jive}{an object of class "jive" (see details)}

\item{log.file}{name of the output file that will store the log of MCMC chain}

\item{sampling.freq}{sampling frequency of the MCMC chain (how often chain will be saved into output file}

\item{print.freq}{printing frequency of the MCMC chain (how often chain will be printed in the R console)}

\item{ncat}{number of classes for thermodynamic integration (see details)}

\item{beta.param}{beta value to define classes for thermodynamic integration (see details)}

\item{ngen}{number of generation in MCMC chain}

\item{burnin}{a burning phase of MCMC chain (has to be specified for thermodynamic integration)}

\item{update.freq}{update frequencies for likelihood and prior level parameters}
}
\value{
none
}
\description{
Implements Markov chain Monte Carlo sampling for trait evolutionary models with intraspecific data
}
\details{
This function runs MCMC sampling on jive object \code{\link{jiveMake}}. The jive object contains
both the dataset and set of model to be used in MCMC. This function implements both a conventional MCMC
and an MCMC with thermodynamic integration. The latter option is turned off by default and can be changed by
setting ncat to values > 1. The recommended ncat for TI is 10. When setting ncat > 1, make sure to specify burning.
As a rule of thumb set burning to 1/10 fraction of ngen.
}
\examples{
## Load test data
data(traitsOU1)
data(treeOU1)
## Run a simple MCMC chain
my.jive <- jiveMake(treeOU1, traitsOU1,  model.var="OU1", model.mean="BM", model.lik="Multinorm")
jiveMCMC(my.jive, log.file="my.jive_MCMC.log", sampling.freq=10, print.freq=100, ngen=5000)

## Run an MCMC chain with thermodynamic integration
jiveMCMC(my.jive, log.file="my.jive_MCMC.log", ncat=10, sampling.freq=10, print.freq=100, ngen=5000, burnin=500)
}
\author{
Anna Kostikova and Daniele Silvestro
}

