% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jivePlot}
\alias{jivePlot}
\title{Plot jive MCMC object}
\usage{
jivePlot(tree, proc.jive, regime, cols = c("blue", "green"),
  cex.label = 0.7, cex.circle = 2, lab.off = 0.025, ladder = TRUE, ...)
}
\arguments{
\item{tree}{an object of class "jive" (see details)}

\item{proc.jive}{name of the output file that will store the log of MCMC chain}

\item{regime}{sampling frequency of the MCMC chain (how often chain will be saved into output file}

\item{cols}{printing frequency of the MCMC chain (how often chain will be printed in the R console)}

\item{cex.label}{number of classes for thermodynamic integration (see details)}

\item{cex.circle}{}

\item{lab.off}{}

\item{ladder}{}
}
\description{
Creates a jive object from intraspecific observations and species phylogeny
}
\details{
This function creates a jive object needed for \code{\link{jiveMCMC}} function.
Trait values must be stored as a matrix, where lines are vector of observations for each species, with NA for no data.
Rownames are species names. Phylogenetic tree must be provided as either simmap object (for models with multiple regimes)
or as a phylo object (for BM or OU1 models)
}
\examples{
jivePlot(phy1, my.l, regime)
}

