% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jiveProc}
\alias{jiveProc}
\title{Process jive MCMC}
\usage{
jiveProc(log.file = "jive_mcmc_OU1.log", n.spec, stat = jiveMode,
  burning = 0, probHPD = 0.95, verbose = TRUE, ...)
}
\arguments{
\item{log.file}{log file recorded by \code{\link{jiveMCMC}} function}

\item{n.spec}{number of species}

\item{stat}{which statistics to use to summarize MCMC. By default, set to mode
for prior level parameters and mean for likelihood level parameters. Can also be \code{\link{mean}}, \code{\link{median}}.}

\item{burning}{how much of burning to disregard}

\item{probHPD}{set HPD intervals}

\item{verbose}{how much of statistics to return}

\item{...}{additional parameters that can be passed to HPDinterval function}
}
\value{
A list of averaged statistics from MCMC chain for each parameter (list)
}
\description{
Process the jiveMCMC output log file
}
\details{
This function processes the output log file of the \code{\link{jiveMCMC}} function.
It summarizes posterior sample for each variable into summary statistics (
e.g. mean, mode, median) and calculates HPD invervals
}
\examples{
## Load test data
data(traitsOU1)
data(treeOU1)
my.jive <- jiveMake(treeOU1, traitsOU1,  model.var="OU1", model.mean="BM", model.lik="Multinorm")
jiveMCMC(my.jive, log.file="OU1_log.log", sampling.freq=10, print.freq=100, ngen=5000)
my.summary <- jiveProc(log.file="OU1_log.log", n = 50, verbose=FALSE)
}
\author{
Anna Kostikova
}

