% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.hpfun}
\alias{make.hpfun}
\title{Title}
\usage{
make.hpfun(hpf = "Uniform", hp.pars, ...)
}
\arguments{
\item{hpf}{description}

\item{hp.pars}{description}

\item{...}{description}
}
\description{
Short description
}
\details{
Detailed description
}
\examples{
my.hp <- make.hpfun(hpf="Uniform", hp.pars=c(1,2))
}

