% Generated by roxygen2 (4.0.1): do not edit by hand
\name{slidingWin}
\alias{slidingWin}
\title{Proposal functions}
\usage{
slidingWin(i, d)
}
\arguments{
\item{i}{fff}

\item{d}{fff}

\item{...}{ff}
}
\description{
Multiplier and sliding window proposals
}
\details{
Make a function to get a hyper prior
}
\examples{
sw <- slidingWin(i=2, d=9)
}

