% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{treeOU1}
\alias{treeOU1}
\title{Jive test dataset: a tree matrix}
\format{Three phylogenetic trees of 50 species in phylo format and three data matrices with 50 rows}
\usage{
treeOU1
}
\description{
A phylogenetic tree of species
}
\details{
This dataset includes a set of simulated trees and trait values. The parameters used for
simulation of these datasets are:
\itemize{
	\item BM. For model.mean - BM with sig.sq = 0.5, anc.state = 350; model.var - BM with sig.sq = , anc.state =
	\item OU1. For model.mean - BM with sig.sq = 0.5, anc.state = 350; model.var - OU1 with alpha = , sig.sq = , anc.state = , theta1 =
	\item BM. For model.mean - BM with sig.sq = 0.5, anc.state = 350; model.var - OUM with alpha = , sig.sq = , anc.state = , theta1 = , theta2 =
}
}
\author{
Anna Kostikova
}
\keyword{data}
\keyword{datasets,}

