% Generated by roxygen2 (4.0.1): do not edit by hand
\name{classicalModel}
\alias{classicalModel}
\title{Simulate trees}
\usage{
classicalModel(nb.divers.shifts, nb.mass.ext, nb.spec, lambda.max, lambda.min,
  mu.min, surv.rate.min, seed = 1)
}
\arguments{
\item{nb.divers.shifts}{number of diversification shifts on the tree}

\item{nb.mass.ext}{number of mass extinction events}

\item{nb.spec}{total number of species in the final tree}

\item{lambda.max}{maximum value of speciation rate}

\item{lambda.min}{minimum value of speciation rate}

\item{mu.min}{minimum value of extinction rate}

\item{surv.rate.min}{survival rate}

\item{seed}{seed value for reproducibility}
}
\value{
An object of class phylo  and a set of parameters used for simulation.
}
\description{
Simulate species trees with complex diversification scenarios
}
\details{
The aim of the function is to simulate species trees with complex diversification scenarios,
with changing diversification rates at inner nodes of the phylogeny and mass extinction events at
particular random times in the past. The user can choose the total number of species, the number
of diversification rate changes, the number of mass extinction events, the minimum and maximum
values for lambda and mu, and the minimum survival rate for mass extinction events. Then the actual rates are choosen randomly.
User gets as output the tree simulated, and all the parameters used for its simulation.
}
\examples{
tree <- classicalModel(1, 1, 500, 0.1, 0.01, .01, .2, seed=1506621)
}
\author{
Sacha Laurent
}

