% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jivePlot}
\alias{jivePlot}
\title{Plot jive MCMC object}
\usage{
jivePlot(tree, proc.jive, regimes, cols = c("blue", "green"),
  cex.label = 0.7, cex.circle = 2, lab.off = 0.025, ladder = TRUE, ...)
}
\arguments{
\item{tree}{an object of class "jive" (see details)}

\item{proc.jive}{an object from \code{\link{jiveProc}} function (see details)}

\item{regimes}{a named vector of selective regimes}

\item{cols}{a vector of colours for selective regimes}

\item{cex.label}{magnification for tip labels}

\item{cex.circle}{magnification for circles representing species-specific variances}

\item{lab.off}{offset for tip labels}

\item{ladder}{if tree should be ladderized}

\item{...}{additional parameters passed to \code{\link{ladderize}} and \code{\link{plot.phylo}} functions}
}
\value{
nothing
}
\description{
Plots jive MCMC output
}
\details{
This function plots estimated species-specific variances on the tree. The spec
The species-specific variances are calculated by \code{\link{jiveProc}}
}
\examples{
## Load test data
data(regimesOU1)
data(traitsOU1)
data(treeOU1)
my.jive <- jiveMake(treeOU1, traitsOU1,  model.var="OU1", model.mean="BM", model.lik="Multinorm")
jiveMCMC(my.jive, log.file="OU1_log.log", sampling.freq=10, print.freq=100, ngen=5000)
my.summary <- jiveProc(log.file="OU1_log.log", n = 50, verbose=FALSE)
jivePlot(treeOU1, my.summary, regimesOU1, cols="red")
}
\author{
Anna Kostikova
}

