% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jivePlot}
\alias{jivePlot}
\title{Plot jive MCMC object}
\usage{
jivePlot(tree, proc.jive, regime, cols = c("blue", "green"),
  cex.label = 0.7, cex.circle = 2, lab.off = 0.025, ladder = TRUE, ...)
}
\arguments{
\item{tree}{an object of class "jive" (see details)}

\item{proc.jive}{an object from \code{\link{jiveProc}} function (see details)}

\item{regime}{a named vector of selective regimes}

\item{cols}{a vector of colours for selective regimes}

\item{cex.label}{magnification for tip labels}

\item{cex.circle}{magnification for circles representing species-specific variances}

\item{lab.off}{offset for tip labels}

\item{ladder}{if tree should be ladderized}

\item{...}{additional parameters passed to \code{\link{ladderize}} and \code{\link{plot.phylo}} functions}
}
\value{
nothing
}
\description{
Plots jive MCMC output
}
\details{
This function plots estimated species-specific variances on the tree. The spec
The species-specific variances are calculated by \code{\link{jiveProc}}
}
\examples{
jivePlot(phy1, my.l, regime)
}
\author{
Anna Kostikova
}

