% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jiveProc}
\alias{jiveProc}
\title{Process jive MCMC}
\usage{
jiveProc(log.file = "jive_mcmc_OU1.log", n = 50, stat = jive.mode,
  burning = 0, probHPD = 0.95, verbose = TRUE, ...)
}
\arguments{
\item{log.file}{log file recorded by jiveMCMC function}

\item{n}{number of species}

\item{stat}{which statistics to use to summarize MCMC. By default set to mode for prior level parameters and mean for likelihood level parameters}

\item{burning}{how much of burning to disregard}

\item{probHPD}{set HPD intervals}

\item{verbose}{how much of statistics to return}

\item{...}{additional parameters that can be passed to HPDinterval function}
}
\description{
What it does
}
\details{
This function processes the output log file of the jiveMCMC function.
It summarizes posterior sample for each variable into summary statistics (
e.g. mean, mode, median) and calculates HPD invervals
}
\examples{
my.summary <- jiveProc(log.file="OU_log.log", n = 50, verbose=FALSE)
}

