% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make.hpfun}
\alias{make.hpfun}
\title{Hyper-prior function}
\usage{
make.hpfun(hpf = "Uniform", hp.pars, ...)
}
\arguments{
\item{hpf}{name of a density function. Supported density functions are: Uniform, Gamma and Normal}

\item{hp.pars}{a vector of density function parameters}

\item{...}{additional parameters that can be passed to a density function}
}
\description{
This function creates a hyper-prior density function.
Currently supported density function are Uniform, Gamma and Normal.
The resulting function is used during MCMC \code{\link{jiveMCMC}}
to estimate parameters of priors.
}
\details{
There are three currently implemented density function:
Uniform, Gamma and Normal. Each of these densities requires two input parameters and hp.pars
must be a vector of two values and cannot be left empty.
}
\examples{
my.hp <- make.hpfun(hpf="Uniform", hp.pars=c(1,2))
}

