% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotTraitgram}
\alias{plotTraitgram}
\title{Plot traitgrams object}
\usage{
plotTraitgram(fitted.model = OUwie_OUM, anc.model = "ER", reps = 5,
  plot.grey.only = "FALSE", ...)
}
\arguments{
\item{fitted.model}{a fitted object of class "OUwie" (see details)}

\item{anc.model}{a name of the model for ancstral state reconstruction (see details)}

\item{reps}{number of replicates for simulating a trait}

\item{plot.grey.only}{if to plot only grey area or more}

\item{...}{additional parameters passed to XXX functions}
}
\value{
nothing
}
\description{
Plots traitgrams under different evolutionary models
}
\details{
This function plots
}
\examples{
## Load test data
library(ape)
library(phytools)
library(OUwie)
tree1		<- rtree(40)
tree1 		<- chronopl(tree1, 1)
regime1		<- rep(c(0,1),each=20)
names(regime1) <- tree1$tip.label
sim1 		<- make.simmap(tree1, regime1, model="ER", nsim=1, pi="estimated")
traits1 <- OUwie.sim(sim1, simmap.tree=TRUE, alpha=c(1.5,1.5), sigma.sq=c(0.1,0.1), theta0=2.5, theta=c(1.0,6.0))
traits1 <- data.frame(traits1[,1],regime1,traits1[,2])
OUwie_OUM <- OUwie(sim1, traits1, model="OUM", simmap.tree=TRUE)
par(mfrow=c(1,2))
plotTraitgram(fitted.model=OUwie_OUM, anc.model="ER", reps=5, plot.grey.only = "TRUE")
plotTraitgram(fitted.model=OUwie_OUM, anc.model="ER", reps=5, plot.grey.only = "FALSE")
}
\author{
Martha Serrano and Daniele Silvestro
}

